Extract `BaseClient`
