+++ etcd3/base_client.py
@@ -1,0 +1,205 @@
+import grpc
+
+import etcd3.exceptions as exceptions
+import etcd3.transactions as transactions
+import etcd3.etcdrpc as etcdrpc
+import etcd3.utils as utils
+
+_EXCEPTIONS_BY_CODE = {
+    grpc.StatusCode.INTERNAL: exceptions.InternalServerError,
+    grpc.StatusCode.UNAVAILABLE: exceptions.ConnectionFailedError,
+    grpc.StatusCode.DEADLINE_EXCEEDED: exceptions.ConnectionTimeoutError,
+    grpc.StatusCode.FAILED_PRECONDITION: exceptions.PreconditionFailedError,
+}
+
+
+def _translate_exception(exc):
+    code = exc.code()
+    exception = _EXCEPTIONS_BY_CODE.get(code)
+    if exception is None:
+        raise
+    raise exception
+
+
+class Transactions(object):
+    def __init__(self):
+        self.value = transactions.Value
+        self.version = transactions.Version
+        self.create = transactions.Create
+        self.mod = transactions.Mod
+
+        self.put = transactions.Put
+        self.get = transactions.Get
+        self.delete = transactions.Delete
+        self.txn = transactions.Txn
+
+
+class KVMetadata(object):
+    def __init__(self, keyvalue, header):
+        self.key = keyvalue.key
+        self.create_revision = keyvalue.create_revision
+        self.mod_revision = keyvalue.mod_revision
+        self.version = keyvalue.version
+        self.lease_id = keyvalue.lease
+        self.response_header = header
+
+
+class Status(object):
+    def __init__(self, version, db_size, leader, raft_index, raft_term):
+        self.version = version
+        self.db_size = db_size
+        self.leader = leader
+        self.raft_index = raft_index
+        self.raft_term = raft_term
+
+
+class Alarm(object):
+    def __init__(self, alarm_type, member_id):
+        self.alarm_type = alarm_type
+        self.member_id = member_id
+
+
+class EtcdTokenCallCredentials(grpc.AuthMetadataPlugin):
+    """Metadata wrapper for raw access token credentials."""
+
+    def __init__(self, access_token):
+        self._access_token = access_token
+
+    def __call__(self, context, callback):
+        metadata = (('token', self._access_token),)
+        callback(metadata, None)
+
+
+class BaseClient(object):
+    @staticmethod
+    def _build_get_range_request(key,
+                                 range_end=None,
+                                 limit=None,
+                                 revision=None,
+                                 sort_order=None,
+                                 sort_target='key',
+                                 serializable=None,
+                                 keys_only=False,
+                                 count_only=None,
+                                 min_mod_revision=None,
+                                 max_mod_revision=None,
+                                 min_create_revision=None,
+                                 max_create_revision=None):
+        range_request = etcdrpc.RangeRequest()
+        range_request.key = utils.to_bytes(key)
+        range_request.keys_only = keys_only
+        if range_end is not None:
+            range_request.range_end = utils.to_bytes(range_end)
+
+        if sort_order is None:
+            range_request.sort_order = etcdrpc.RangeRequest.NONE
+        elif sort_order == 'ascend':
+            range_request.sort_order = etcdrpc.RangeRequest.ASCEND
+        elif sort_order == 'descend':
+            range_request.sort_order = etcdrpc.RangeRequest.DESCEND
+        else:
+            raise ValueError('unknown sort order: "{}"'.format(sort_order))
+
+        if sort_target is None or sort_target == 'key':
+            range_request.sort_target = etcdrpc.RangeRequest.KEY
+        elif sort_target == 'version':
+            range_request.sort_target = etcdrpc.RangeRequest.VERSION
+        elif sort_target == 'create':
+            range_request.sort_target = etcdrpc.RangeRequest.CREATE
+        elif sort_target == 'mod':
+            range_request.sort_target = etcdrpc.RangeRequest.MOD
+        elif sort_target == 'value':
+            range_request.sort_target = etcdrpc.RangeRequest.VALUE
+        else:
+            raise ValueError('sort_target must be one of "key", '
+                             '"version", "create", "mod" or "value"')
+
+        return range_request
+
+    @staticmethod
+    def _build_put_request(key, value, lease=None, prev_kv=False):
+        put_request = etcdrpc.PutRequest()
+        put_request.key = utils.to_bytes(key)
+        put_request.value = utils.to_bytes(value)
+        put_request.lease = utils.lease_to_id(lease)
+        put_request.prev_kv = prev_kv
+
+        return put_request
+
+    @staticmethod
+    def _build_delete_request(key,
+                              range_end=None,
+                              prev_kv=False):
+        delete_request = etcdrpc.DeleteRangeRequest()
+        delete_request.key = utils.to_bytes(key)
+        delete_request.prev_kv = prev_kv
+
+        if range_end is not None:
+            delete_request.range_end = utils.to_bytes(range_end)
+
+        return delete_request
+
+    def _ops_to_requests(self, ops):
+        """
+        Return a list of grpc requests.
+
+        Returns list from an input list of etcd3.transactions.{Put, Get,
+        Delete, Txn} objects.
+        """
+        request_ops = []
+        for op in ops:
+            if isinstance(op, transactions.Put):
+                request = self._build_put_request(op.key, op.value,
+                                                  op.lease, op.prev_kv)
+                request_op = etcdrpc.RequestOp(request_put=request)
+                request_ops.append(request_op)
+
+            elif isinstance(op, transactions.Get):
+                request = self._build_get_range_request(op.key, op.range_end)
+                request_op = etcdrpc.RequestOp(request_range=request)
+                request_ops.append(request_op)
+
+            elif isinstance(op, transactions.Delete):
+                request = self._build_delete_request(op.key, op.range_end,
+                                                     op.prev_kv)
+                request_op = etcdrpc.RequestOp(request_delete_range=request)
+                request_ops.append(request_op)
+
+            elif isinstance(op, transactions.Txn):
+                compare = [c.build_message() for c in op.compare]
+                success_ops = self._ops_to_requests(op.success)
+                failure_ops = self._ops_to_requests(op.failure)
+                request = etcdrpc.TxnRequest(compare=compare,
+                                             success=success_ops,
+                                             failure=failure_ops)
+                request_op = etcdrpc.RequestOp(request_txn=request)
+                request_ops.append(request_op)
+
+            else:
+                raise Exception(
+                    'Unknown request class {}'.format(op.__class__))
+        return request_ops
+
+    @staticmethod
+    def _build_alarm_request(alarm_action, member_id, alarm_type):
+        alarm_request = etcdrpc.AlarmRequest()
+
+        if alarm_action == 'get':
+            alarm_request.action = etcdrpc.AlarmRequest.GET
+        elif alarm_action == 'activate':
+            alarm_request.action = etcdrpc.AlarmRequest.ACTIVATE
+        elif alarm_action == 'deactivate':
+            alarm_request.action = etcdrpc.AlarmRequest.DEACTIVATE
+        else:
+            raise ValueError('Unknown alarm action: {}'.format(alarm_action))
+
+        alarm_request.memberID = member_id
+
+        if alarm_type == 'none':
+            alarm_request.alarm = etcdrpc.NONE
+        elif alarm_type == 'no space':
+            alarm_request.alarm = etcdrpc.NOSPACE
+        else:
+            raise ValueError('Unknown alarm type: {}'.format(alarm_type))
+
+        return alarm_request
